+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [Provide an Oauth2 scope for the staff role #127](https://github.com/farmier/farm_rcd/issues/127)
 - [Enable staff to comment on plans #126](https://github.com/farmier/farm_rcd/issues/126)
+- [View of plans by intake #125](https://github.com/farmier/farm_rcd/issues/125)
 
 ## [1.2.0] 2026-01-21